add port I/O functions
Add datatype aliases

+++ utils/mem.h
@@ -1,13 +1,12 @@
 #ifndef _UTILS_MEM
 #define _UTILS_MEM
 
-#include <stdint.h>
-#include <stddef.h>
+#include "../cpu/types.h"
 
-void memcpy(uint8_t *source, uint8_t *dest, int nbytes);
-void memset(uint8_t *dest, uint8_t val, uint32_t len);
+void memcpy(u8 *source, u8 *dest, int nbytes);
+void memset(u8 *dest, u8 val, u32 len);
 
 /* At this stage there is no 'free' implemented. */
-uint32_t kmalloc(size_t size, int align, uint32_t *phys_addr);
+u32 kmalloc(int size, int align, u32 *phys_addr);
 
 #endif
